% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-pattern-centers.R
\name{center_median}
\alias{center_median}
\title{Calculate Median Center}
\usage{
center_median(geometry)
}
\arguments{
\item{geometry}{an sfc object. If a polygon, uses \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}.}
}
\description{
Calculate Median Center
}
\seealso{
Other point-pattern: 
\code{\link{center_mean}()}
}
\concept{point-pattern}
