% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_suggests}
\alias{check_pkg_suggests}
\title{Check if a vector of packages are available}
\usage{
check_pkg_suggests(x)
}
\arguments{
\item{x}{a character vector of package names}
}
\description{
Check if a vector of packages are available
}
\keyword{internal}
