% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_modify}
\alias{class_modify}
\title{Modify object classes}
\usage{
class_modify(x, class = "list")
}
\arguments{
\item{x}{an object to modify}

\item{class}{a character vector of classes to append to an object}
}
\description{
Appends classes to exist object classes. This is utilized to aid in adding a list class to objects created by spdep. This enables to use of the returned objects within data frames and tibbles.
}
\keyword{internal}
