% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emerging-hostpot-classifications.R
\name{classify_hotspot}
\alias{classify_hotspot}
\title{Classify Hot Spot results}
\usage{
classify_hotspot(.x, threshold)
}
\description{
Given the Gi* time-series and Mann Kendall scores classify the hotspot values
}
\keyword{internal}
