% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-pattern-centers.R
\name{euclidean_median}
\alias{euclidean_median}
\title{Calculate Euclidean Median Center}
\usage{
euclidean_median(geometry, tolerance = 1e-09)
}
\arguments{
\item{geometry}{an sfc object. If a polygon, uses \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}.}

\item{tolerance}{a tolerance level to terminate the process. This is passed to \code{\link[pracma:geo_median]{pracma::geo_median()}}.}
}
\description{
Given an sfc geometry, calculate the Euclidean Median Center.
}
\details{
Calculation of the Euclidean median is done using the pracma package. If Rfast is not available, the function will error.
}
