% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmt-impl.R
\name{folded_p_sim}
\alias{folded_p_sim}
\title{Folded Simulated P-value}
\usage{
folded_p_sim(obs, reps)
}
\arguments{
\item{obs}{observed values}

\item{reps}{a matrix}
}
\description{
Given observed statistics and replicates, calculated a simulated folded p value.
}
