% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-colocation-impl.R
\name{global_colocation_calc}
\alias{global_colocation_calc}
\title{Calculate the Global Colocation Quotient}
\usage{
global_colocation_calc(A, nb)
}
\description{
Given a categorical variable \code{A}, calculate the global colocation quotient (CLQ).
}
\keyword{internal}
