% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-colocation-impl.R
\name{global_colocation_perm_impl}
\alias{global_colocation_perm_impl}
\title{Global Colocation Quotient Conditional Permutation Implementation}
\usage{
global_colocation_perm_impl(A, nb, nsim = 99)
}
\description{
This function will repeat calculation of the global colocation quotient (CLQ)
a specified number of times and return a simulated p-value.
}
\keyword{internal}
