% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-measures.R
\name{global_g_test}
\alias{global_g_test}
\title{Getis-Ord Global G}
\usage{
global_g_test(x, nb, wt, alternative = "greater", allow_zero = NULL, ...)
}
\description{
Getis-Ord Global G
}
\examples{
global_g_test(x, nb, wt)
}
