% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-measures.R
\name{global_jc_perm}
\alias{global_jc_perm}
\alias{global_jc_test}
\title{Global Join Count Permutation Test}
\usage{
global_jc_perm(
  fx,
  nb,
  wt,
  alternative = "greater",
  nsim = 499,
  allow_zero = FALSE,
  ...
)

global_jc_test(fx, nb, wt, alternative = "greater", allow_zero = NULL, ...)
}
\arguments{
\item{fx}{a factor or character vector of the same length as nb.}

\item{nb}{a neighbor list object for example as created by \code{st_contiguity()}.}

\item{wt}{a weights list as created by \code{st_weights()}.}

\item{nsim}{number of simulations to run.}

\item{...}{additional arguments passed to \code{\link[spdep:joincount.test]{spdep::joincount.test()}}}
}
\description{
Global Join Count Permutation Test

Global Join Count Test
}
\examples{
fx <- guerry$region
global_jc_perm(fx, nb, wt)
global_jc_test(fx, nb, wt)
}
