% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guerry}
\alias{guerry}
\alias{guerry_nb}
\title{"Essay on the Moral Statistics of France" data set.}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 85 rows and 27 columns.

\code{guerry} an sf object with 85 observations and 27 variables. \code{guerry_nb} has 2 additional variables created by \code{sfdep}.
}
\source{
\code{Guerry::gfrance85}
}
\usage{
guerry

guerry_nb
}
\description{
This dataset has been widely used to demonstrate geospatial methods and techniques. As such it is useful for inclusion to this R package for the purposes of example. The dataset in this package is modified from Guerry by \href{https://www.datavis.ca/}{Michael Friendly}.
}
\details{
\code{guerry} and \code{guerry_nb} objects are sf class objects. These are polygons of the boundaries of France (excluding Corsica) as they were in 1830.
}
\keyword{datasets}
