% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-self.R
\name{include_self}
\alias{include_self}
\alias{remove_self}
\title{Includes self in neighbor list}
\usage{
include_self(nb)

remove_self(nb)
}
\arguments{
\item{nb}{an object of class \code{nb} e.g. made by \code{\link[=st_contiguity]{st_contiguity()}}}
}
\value{
An object of class \code{nb}.
}
\description{
Includes observed region in list of own neighbors. For some neighbor lists, it is important to include the ith observation (or self) in the neighbors list, particulalry for kernel weights.
}
\examples{
nb <- st_contiguity(guerry)
self_included <- include_self(nb)
self_included
remove_self(self_included)
}
