% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights-inverse-distance.R
\name{inverse_dist_calc}
\alias{inverse_dist_calc}
\title{Calculate inverse distance weights}
\usage{
inverse_dist_calc(.x, scale, alpha)
}
\description{
Calculate inverse distance weights
}
\keyword{internal}
