% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount-bv-impl.R
\name{jc_bjc_perm_impl}
\alias{jc_bjc_perm_impl}
\title{Calculate BJC BV for conditional permutations}
\usage{
jc_bjc_perm_impl(x, z, listw, index)
}
\description{
Calculate BJC BV for conditional permutations
}
\keyword{internal}
