% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount-bv-impl.R
\name{jc_clc_perm_impl}
\alias{jc_clc_perm_impl}
\title{Calculate CLC BV for conditional permutations}
\usage{
jc_clc_perm_impl(x, z, listw, index)
}
\arguments{
\item{x}{a binary variable consisting of 1 and 0, or \code{TRUE} and \code{FALSE}.}

\item{z}{a binary variable consisting of 1 and 0, or \code{TRUE} and \code{FALSE}.}

\item{listw}{a \code{listw} object}

\item{index}{an integer vector identifying positions to subset.}
}
\description{
Calculate CLC BV for conditional permutations
}
\keyword{internal}
