% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount-uni-impl.R
\name{jc_uni_calc}
\alias{jc_uni_calc}
\title{Calculate univariate join count at row level}
\usage{
jc_uni_calc(x, xj, wt)
}
\description{
Calculates the local join count statistic for a given xi, xj, and binary weights.
}
\details{
Formula is

\eqn{x_i \times \Sigma{w_i * x_j}}
}
\keyword{internal}
