% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount-uni-impl.R
\name{jc_uni_perm_impl}
\alias{jc_uni_perm_impl}
\title{Conditional permutation of univariate local join count}
\usage{
jc_uni_perm_impl(x, listw, index)
}
\arguments{
\item{x}{binary variable}

\item{listw}{must be binary}

\item{index}{to subset}
}
\description{
Creates conditional permutations on an index and calculates the local univariate
join count for that index.
}
\details{
TODO this conditional permutation indexing can be useful elsewhere if subsetting
is further required for other functions
}
\keyword{internal}
