% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-colocation-quotient-impl.R
\name{local_colocation}
\alias{local_colocation}
\title{Local Colocation Quotient}
\usage{
local_colocation(A, B, nb, wt, nsim)
}
\arguments{
\item{A}{a character or factor vector.}

\item{B}{a character or factor vector.}

\item{nb}{a neighbors list e.g. created by \code{\link[=st_knn]{st_knn()}} or \code{\link[=st_contiguity]{st_contiguity()}}}

\item{nsim}{default \code{99}. An integer representing how many simulations to run for calculating the simulated p-values.}
}
\value{
a data frame with as many rows as observations in A and two times as many columns as unique values in B.
}
\description{
The Local Colocation Quotient requires guassian weights.
Guassian weights help by adding a \emph{decay} to the relationship between points
Take point xi as the focal point and xij1 and xij2 as neighbors. xij1 is 10 units  and xij2 is 15 units away respectively. The influence of xij1 is greater than xij2 due to their proximity.
}
\details{
Uses an adaptive bandwidth

Could use network distance. Less efficient in calculation than euclidian but more intuitive or informative
}
\examples{
A <- guerry$main_city
B <- guerry$region
geo <- sf::st_geometry(guerry)
nb <- st_knn(geo, 5)
wt <- st_kernel_weights(nb, geo, "gaussian")
res <- local_colocation(A, B, nb, wt, 49)
head(res)
}
