% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-colocation-quotient-impl.R
\name{local_colocation_calc}
\alias{local_colocation_calc}
\title{Calculate the local colocation quotient}
\usage{
local_colocation_calc(A, B, listw)
}
\value{
a matrix
}
\description{
Calculate the local colocation quotient
}
\keyword{internal}
