% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-colocation-quotient-impl.R
\name{local_colocation_impl}
\alias{local_colocation_impl}
\title{spdep implementation of local colocation quotient}
\usage{
local_colocation_impl(A, B, listw, nsim = 99)
}
\value{
a data frame where the number of rows is the same length as \code{A} and the number of columns is the same as unique values in \code{B}.
}
\description{
Internal implementation of the local CLQ that is compatible with spdep.
}
\keyword{internal}
