% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-local-gistar-impl.R
\name{local_g_spt}
\alias{local_g_spt}
\title{Calculate the Local Gi* for a spacetime cube}
\usage{
local_g_spt(x, times, nb, wt, n_locs, nsim)
}
\arguments{
\item{x}{a numeric vector}

\item{times}{a vector determining time order}

\item{nb}{a spacetime neighbors list}

\item{wt}{a weights list associated with \code{nb}}

\item{n_locs}{the number of unique locations}

\item{n_sim}{the number of simulations to run for calculating \code{p_sim}}
}
\description{
An alternative implementation to \link[spdep:localG]{spdep::localG} intended for use with
a spacetime cube.
}
\keyword{internal}
