% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-moran-bv-impl.R
\name{local_moran_bv_perm_impl}
\alias{local_moran_bv_perm_impl}
\title{Local Bivariate Moran's I conditional permutation implementation}
\usage{
local_moran_bv_perm_impl(x, y, listw)
}
\arguments{
\item{x}{a numeric vector of same length as \code{y}.}

\item{y}{a numeric vector of same length as \code{x}.}

\item{listw}{a listw object for example as created by \code{nb2listw()}.}
}
\description{
Local Bivariate Moran's I conditional permutation implementation
}
\keyword{internal}
