% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-moran-bv-impl.R
\name{moran_bv_calc}
\alias{moran_bv_calc}
\title{Global Bivariate Moran's I calculation}
\usage{
moran_bv_calc(x, yj, wt)
}
\arguments{
\item{x}{a numeric vector of same length as \code{nb}.}

\item{yj}{neighbor values of \code{y}, same length and lengths as \code{nb}.}

\item{wt}{a weights list as created by \code{st_weights()}.}
}
\description{
Global Bivariate Moran's I calculation
}
\keyword{internal}
