% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-moran-bv-impl.R
\name{moran_bv_perm_impl}
\alias{moran_bv_perm_impl}
\title{Global Bivariate Moran's I conditional permutation implementation}
\usage{
moran_bv_perm_impl(x, y, listw)
}
\arguments{
\item{x}{a numeric vector of same length as \code{nb}.}

\item{y}{a numeric vector of same length as \code{nb}.}

\item{listw}{a listw object e.g. as created by \code{\link[=recreate_listw]{recreate_listw()}}.}
}
\description{
Global Bivariate Moran's I conditional permutation implementation
}
\keyword{internal}
