% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb-match-test-impl.R
\name{nmt_calc}
\alias{nmt_calc}
\title{Identify matches between two neighbor lists}
\usage{
nmt_calc(knn_nb, nb)
}
\arguments{
\item{knn_nb}{a list with the same length as number of observations in the used dataset}

\item{nb}{a list the same length as knn_nb.}
}
\description{
Identify matches between two neighbor lists
}
\keyword{internal}
