% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pariwise-colocation-impl.R
\name{pairwise_colocation_perm_impl}
\alias{pairwise_colocation_perm_impl}
\title{Pairwise CLQ conditional permutation implementation}
\usage{
pairwise_colocation_perm_impl(A, B, nb, nsim = 199)
}
\arguments{
\item{A}{a character or factor vector.}

\item{B}{a character or factor vector.}

\item{nb}{a neighbors list e.g. created by \code{\link[=st_knn]{st_knn()}} or \code{\link[=st_contiguity]{st_contiguity()}}}

\item{nsim}{default \code{99}. An integer representing how many simulations to run for calculating the simulated p-values.}
}
\description{
Internal implementation of the pairwise CQL using conditional permutation. Functionally the same as \code{\link[=pairwise_colocation]{pairwise_colocation()}}.
}
\keyword{internal}
