% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untils-constants.R
\name{pct_nonzero}
\alias{pct_nonzero}
\title{Percent Non-zero Neighbors}
\usage{
pct_nonzero(nb)
}
\arguments{
\item{nb}{a neighbor list object}
}
\value{
a scalar double
}
\description{
Calculate the percentage of non-zero neighbors in a neighbor list.
}
