% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_listw}
\alias{permute_listw}
\title{Conditionally permutes a listw object}
\usage{
permute_listw(listw)
}
\arguments{
\item{a}{listw object.}
}
\description{
Conditionally permutes a listw object
}
\keyword{internal}
