% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffle_nbs}
\alias{shuffle_nbs}
\title{Internal function to shuffle neighbors}
\usage{
shuffle_nbs(i, n, card)
}
\arguments{
\item{i}{the index position of observation \code{i}}
}
\value{
a \code{nb} class object
}
\description{
Used in conditional permutation and the function \code{permute_listw()}.
}
\keyword{internal}
