% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspace-trig.R
\name{sin_d}
\alias{sin_d}
\title{Trigonometric functions}
\usage{
sin_d(theta = 0)
}
\description{
These functions are wrappers around the base trigonometric functions.
However these return degree values rather than radians.
}
\details{
These functions were adopted from the now archived CRAN package \code{{aspace}}.
}
\keyword{internal}
