% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-cube-set-cols.R
\name{spt_order}
\alias{spt_order}
\title{Order a spacetime cube}
\usage{
spt_order(x)
}
\arguments{
\item{x}{a spacetime cube object.}
}
\description{
When performing analysis on a spacetime cube, the order is of the utmost
importance. This function ensures that a spacetime cube is ordered
appropriately.
}
\keyword{internal}
