% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-methods.R
\name{spt_update}
\alias{spt_update}
\title{Update spacetime attributes}
\usage{
spt_update(x, ...)
}
\arguments{
\item{x}{a spacetime object}

\item{...}{unused}
}
\description{
Update's a spacetime object's number of locations and time periods.
A spacetime object's attributes are sticky and will not change if subsetted
for example by using \code{\link[dplyr:filter]{dplyr::filter()}} or \code{\link[dplyr:slice]{dplyr::slice()}}. Update the
locations and times of a spacetime object.
}
