% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetworks.R
\name{st_as_edges}
\alias{st_as_edges}
\alias{st_as_edges.sf}
\alias{st_as_edges.sfc}
\title{Convert to an edge lines object}
\usage{
st_as_edges(x, ...)

\method{st_as_edges}{sf}(x, nb, wt)

\method{st_as_edges}{sfc}(x, nb, wt)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}.}

\item{nb}{a neighbor list. If \code{x} is class \code{sf}, the unquote named of the column. If \code{x} is class \code{sfc}, an object of class \code{nb} as created from \code{st_contiguity()}.}

\item{wt}{a weights list. If \code{x} is class \code{sf}, the unquote named of the column. If \code{x} is class \code{sfc}, the weights list itself.}
}
\description{
Given geometry and neighbor and weights lists, create an edge list \code{sf} object.
}
