% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb-complete-graph.R
\name{st_complete_nb}
\alias{st_complete_nb}
\title{Create Neighbors as Complete Graph}
\usage{
st_complete_nb(n_elements, diag = FALSE)
}
\arguments{
\item{n_elements}{the number of observations to create a neighbors list for}

\item{diag}{default \code{FALSE}. If \code{TRUE}, includes diagonal element / the self.}
}
\value{
A neighbors list representing a complete graph.
}
\description{
Create a neighbors list where every element is related to every other element.
This creates a complete graph.
}
\examples{
st_complete_nb(5)
}
