% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-methods.R
\name{[.spacetime}
\alias{[.spacetime}
\title{TODO update \code{times} attribute when subset and data is active}
\usage{
\method{[}{spacetime}(x, ...)
}
\description{
TODO update \code{times} attribute when subset and data is active
}
