% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-constants.R
\name{szero}
\alias{szero}
\title{Global sum of weights}
\usage{
szero(wt)
}
\arguments{
\item{wt}{a weights list—i.e. created by \code{st_weights()}}
}
\value{
a scalar numeric
}
\description{
Calculate the global sum of weights
}
\examples{
nb <- st_contiguity(guerry)
wt <- st_weights(nb)
szero(wt)
}
