% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-methods.R
\name{tidyverse}
\alias{tidyverse}
\alias{group_by.spacetime}
\alias{mutate.spacetime}
\title{tidyverse methods for spacetime objects}
\usage{
group_by.spacetime(.data, ...)

mutate.spacetime(.data, ...)
}
\description{
dplyr methods for spacetime objects.
}
