% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-methods.R
\name{as_sf}
\alias{as_sf}
\alias{as_spacetime}
\alias{as_spacetime.sf}
\title{Cast between \code{spacetime} and \code{sf} classes}
\usage{
as_sf(x, ...)

as_spacetime(x, .loc_col, .time_col, ...)

\method{as_spacetime}{sf}(x, .loc_col, .time_col, ...)
}
\arguments{
\item{x}{for \code{\link[=st_as_sf]{st_as_sf()}} a spacetime object. For \code{\link[=as_spacetime]{as_spacetime()}}
an sf object.}

\item{...}{unused}

\item{.loc_col}{the quoted name of the column containing unique location identifiers.}

\item{.time_col}{the quoted name of the column containing time periods.}
}
\description{
Cast between \code{spacetime} and \code{sf} classes

Convert sf object to spacetime
}
