% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-pattern-centers.R
\name{center_mean}
\alias{center_mean}
\title{Calculate Center Mean Point}
\usage{
center_mean(geometry, weights = NULL)
}
\arguments{
\item{geometry}{an sfc object. If a polygon, uses \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}.}

\item{weights}{an optional vector of weights to apply to the coordinates before calculation.}
}
\description{
Given an sfc object, calculates the mean center or weighted mean center of the points.
}
\seealso{
Other point-pattern: 
\code{\link{center_median}()}
}
\concept{point-pattern}
