% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-cube-complete.R
\name{complete_spacetime_cube}
\alias{complete_spacetime_cube}
\title{Convert spacetime object to spacetime cube}
\usage{
complete_spacetime_cube(x, ...)
}
\arguments{
\item{x}{a spacetime object.}

\item{...}{unused}
}
\description{
Given a spacetime object, convert it to a spacetime cube. A spacetime cube
ensures that there is a regular time-series for each geometry present.
}
\details{
If observations are missing for a time period and location combination,
columns will be populated with NAs.

See \code{\link[=is_spacetime_cube]{is_spacetime_cube()}} for more details on spacetime cubes.
}
