% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-measures.R
\name{global_c}
\alias{global_c}
\title{Compute Geary's C}
\usage{
global_c(x, nb, wt, allow_zero = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{nb}{a neighbor list object for example as created by \code{st_contiguity()}.}

\item{wt}{a weights list as created by \code{st_weights()}.}

\item{allow_zero}{If \code{TRUE}, assigns zero as lagged value to zone without neighbors.}
}
\value{
a list with two names elements \code{C} and \code{K} returning the value of Geary's C and sample kurtosis respectively.
}
\description{
Compute Geary's C
}
\examples{
nb <- guerry_nb$nb
wt <- guerry_nb$wt
x <- guerry_nb$crime_pers
global_c(x, nb, wt)
}
\seealso{
Other global_c: 
\code{\link{global_c_perm}()},
\code{\link{global_c_test}()}
}
\concept{global_c}
