% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-measures.R
\name{global_c_perm}
\alias{global_c_perm}
\title{Global C Permutation Test}
\usage{
global_c_perm(
  x,
  nb,
  wt,
  nsim = 499,
  alternative = "greater",
  allow_zero = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{nb}{a neighbor list object for example as created by \code{st_contiguity()}.}

\item{wt}{a weights list as created by \code{st_weights()}.}

\item{nsim}{number of simulations to run.}

\item{...}{additional arguments passed to \code{\link[spdep:geary.mc]{spdep::geary.mc()}}.}
}
\description{
Global C Permutation Test
}
\examples{
global_c_perm(x, nb, wt)
}
\seealso{
Other global_c: 
\code{\link{global_c_test}()},
\code{\link{global_c}()}
}
\concept{global_c}
