% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime-cube-test.R
\name{is_spacetime_cube}
\alias{is_spacetime_cube}
\title{Test if a spacetime object is a spacetime cube}
\usage{
is_spacetime_cube(x, ...)
}
\arguments{
\item{x}{a spacetime object}

\item{...}{unused}
}
\description{
Given an object with class \code{spacetime}, determine if it is a \emph{spacetime cube}.
}
