% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{kernels}
\alias{kernels}
\title{Match kernel functions}
\format{
An object of class \code{list} of length 5.
}
\usage{
kernels
}
\arguments{
\item{x}{an object}

\item{thresh}{threshold}

\item{...}{unused}
}
\description{
Match kernel functions
}
\keyword{datasets}
