% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localG.R
\name{local_g}
\alias{local_g}
\alias{local_g_perm}
\title{Local G}
\usage{
local_g(x, nb, wt, alternative = "two.sided", ...)

local_g_perm(x, nb, wt, nsim = 499, alternative = "two.sided", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{nb}{a neighbor list object for example as created by \code{st_contiguity()}.}

\item{wt}{a weights list as created by \code{st_weights()}.}

\item{...}{methods passed to \code{\link[spdep:localG]{spdep::localG()}} or \code{\link[spdep:localG]{spdep::localG_perm()}}
x <- guerry$crime_pers
nb <- st_contiguity(guerry)
wt <- st_weights(nb)}

\item{nsim}{The number of simulations to run.}
}
\description{
Local G
}
