% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_moran}
\alias{local_moran}
\title{Calculate the Local Moran's I Statistic}
\usage{
local_moran(x, nb, wt, alternative = "two.sided", nsim = 499, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{nsim}{The number of simulations to run.}

\item{...}{See \code{?spdep::localmoran_perm()} for more options.}
}
\value{

}
\description{
Moran's I is calculated for each polygon based on the neighbor and weight lists.
}
\details{
\code{\link[=local_moran]{local_moran()}} calls \code{\link[spdep:localmoran]{spdep::localmoran_perm()}} and calculates the Moran I for each polygon. As well as provide simulated p-values.
}
\examples{
library(tidyverse)

lisa <- guerry \%>\%
  mutate(nb = st_contiguity(geometry),
         wt = st_weights(nb),
         moran = local_moran(crime_pers, nb, wt))

# unnest the dataframe column
tidyr::unnest(lisa, moran)
}
\seealso{
Other stats: 
\code{\link{st_lag}()},
\code{\link{st_nb_dists}()}
}
\concept{stats}
