% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_moran}
\alias{local_moran}
\title{Calculate the Local Moran's I Statistic}
\usage{
local_moran(
  x,
  nb,
  wt,
  alternative = "two.sided",
  nsim = 499,
  conditional = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{See \code{?spdep::localmoran()} for more options.}
}
\description{
Calculate the Local Moran's I Statistic
}
\examples{
library(tidyverse)


pluck(lisa, "moran")
}
\seealso{
Other stats: 
\code{\link{st_lag}()},
\code{\link{st_nb_dists}()}
}
\concept{stats}
