% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-moran-bv-impl.R
\name{local_moran_bv_calc}
\alias{local_moran_bv_calc}
\title{Calculate the Local Bivariate Moran Statistic}
\usage{
local_moran_bv_calc(x, y, nb, wt)
}
\arguments{
\item{x}{a numeric vector of same length as \code{y}.}

\item{y}{a numeric vector of same length as \code{x}.}
}
\description{
Calculate the Local Bivariate Moran Statistic
}
\keyword{internal}
