% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb-match-test-impl.R
\name{nmt_perm_impl}
\alias{nmt_perm_impl}
\title{Find conditionally permuted neighbor matches
Given a kNN attribute neighbor list and a listw object, find the number of matches given a conditional permutation.}
\usage{
nmt_perm_impl(knn_nb, listw)
}
\arguments{
\item{knn_nb}{a list with numeric elements. For example as made by \code{\link[dbscan:NN]{dbscan::adjacencylist()}}}

\item{listw}{a listw object likely created by \link{recreate_listw}.}
}
\description{
Find conditionally permuted neighbor matches
Given a kNN attribute neighbor list and a listw object, find the number of matches given a conditional permutation.
}
\keyword{internal}
