% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetworks.R
\name{st_as_nodes}
\alias{st_as_nodes}
\alias{st_as_nodes.sf}
\alias{st_as_nodes.sfc}
\title{Convert to a node point object}
\usage{
st_as_nodes(x, ...)

\method{st_as_nodes}{sf}(x, nb)

\method{st_as_nodes}{sfc}(x, nb)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}.}

\item{nb}{a neighbor list. If \code{x} is class \code{sf}, the unquote named of the column. If \code{x} is class \code{sfc}, an object of class \code{nb} as created from \code{st_contiguity()}.}
}
\description{
Given geometry and a neighbor list, creates an \code{sf} object to be used as nodes in an \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork()}}.
}
