% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{st_contiguity}
\alias{st_contiguity}
\title{Identify polygon neighbors}
\usage{
st_contiguity(x, queen = TRUE, ...)
}
\arguments{
\item{x}{an sf or sfc object.}

\item{queen}{Default \code{TRUE}. For more see \code{?spdep::poly2nb}}
}
\description{
Identify polygon neighbors
}
\examples{
# on basic polygons
geo <- sf::st_geometry(guerry)
st_contiguity(geo)

# in a pipe
guerry \%>\%
  dplyr::mutate(nb = st_contiguity(geometry), .before = 1)
}
\seealso{
Other neighbors: 
\code{\link{st_knn}()},
\code{\link{st_nb_band}()}
}
\concept{neighbors}
