% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{st_inverse_weights}
\alias{st_inverse_weights}
\title{Calculate Inverse Distance Bands}
\usage{
st_inverse_weights(x, nb, scale = 100, threshold = NULL)
}
\arguments{
\item{x}{Spatial points. Typically the \code{geometry} column of an sf object.}

\item{nb}{A nb neighbor list such as created from \code{st_knn(x, k = 1)}.}

\item{scale}{The scaling factor to use when calculating the inverse distance band.}

\item{threshold}{The critical threshold to use for distance band.}
}
\description{
Calculate Inverse Distance Bands
}
\details{
See implementation details \href{https://spatialanalysis.github.io/lab_tutorials/Spatial_Weights_as_Distance_Functions.html#kernal-weights}{here}. For more on distance band based weights see \href{https://spatialanalysis.github.io/lab_tutorials/Distance_Based_Spatial_Weights.html#creating-distance-band-weights}{here}.
}
\seealso{
Other weights: 
\code{\link{st_kernel_weights}()},
\code{\link{st_weights}()}
}
\concept{weights}
