% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{st_kernel_weights}
\alias{st_kernel_weights}
\title{Calculate Kernel Weights}
\usage{
st_kernel_weights(x, nb, kernel = "uniform", threshold = NULL)
}
\arguments{
\item{x}{Spatial points. Typically the \code{geometry} column of an sf object.}

\item{nb}{A nb neighbor list such as created from \code{st_knn(x, k = 1)}.}

\item{kernel}{One of "uniform", "gaussian",  "triangular", "epanechnikov", or "quartic".}

\item{threshold}{The critical threshold to use for distance band.}
}
\description{
Calculate Kernel Weights
}
\details{
By default \code{st_kernel_weight()} utilizes a critical threshold of the maximum neighbor distance. If desired, the critical threshold can be specified manually. The \code{threshold} will be passed to the underlying kernel.

See kernels for more.
}
\seealso{
Other weights: 
\code{\link{st_inverse_weights}()},
\code{\link{st_weights}()}
}
\concept{weights}
