% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{st_knn}
\alias{st_knn}
\title{Calculate K-Nearest Neighbors}
\usage{
st_knn(x, k = 1, symmetric = FALSE, ...)
}
\arguments{
\item{x}{an sf or sfc object.}

\item{k}{number of nearest neighbours to be returned}

\item{symmetric}{default \code{FALSE}. Whether to force output of neighbours to be symmetric.}

\item{...}{additional arguments to be passed to \code{knearneigh()}.}
}
\description{
Calculate K-Nearest Neighbors
}
\seealso{
Other neighbors: 
\code{\link{st_contiguity}()},
\code{\link{st_nb_band}()}
}
\concept{neighbors}
