% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{st_weights}
\alias{st_weights}
\title{Calculate spatial weights}
\usage{
st_weights(nb, style = "W", allow_zero = NULL, ...)
}
\arguments{
\item{nb}{A neighbor list object as created by \code{st_neighbors()}.}

\item{style}{Default \code{"W"} for row standardized weights.}
}
\description{
Calculate spatial weights
}
\seealso{
Other weights: 
\code{\link{st_inverse_weights}()},
\code{\link{st_kernel_weight}()}
}
\concept{weights}
