% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-pattern_std-distance.R
\name{std_distance}
\alias{std_distance}
\title{Calculate standard distance}
\usage{
std_distance(geometry)
}
\arguments{
\item{geometry}{an sfc object. If a polygon, uses \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}.}
}
\description{
Calculate standard distance
}
